LayoutTests/http/tests/adClickAttribution/anchor-tag-attributes-validation-expected.txt

CONSOLE MESSAGE: line 108: adcampaignid must have a non-negative value less than 64 for Ad Click Attribution.
CONSOLE MESSAGE: line 108: adcampaignid must have a non-negative value less than 64 for Ad Click Attribution.
CONSOLE MESSAGE: line 108: adcampaignid can not be converted to a non-negative integer which is required for Ad Click Attribution.
CONSOLE MESSAGE: line 108: adcampaignid can not be converted to a non-negative integer which is required for Ad Click Attribution.
CONSOLE MESSAGE: line 108: adcampaignid can not be converted to a non-negative integer which is required for Ad Click Attribution.
CONSOLE MESSAGE: line 108: addestination could not be converted to a valid HTTP-family URL.
CONSOLE MESSAGE: line 108: addestination could not be converted to a valid HTTP-family URL.
CONSOLE MESSAGE: line 108: addestination could not be converted to a valid HTTP-family URL.
CONSOLE MESSAGE: line 108: Both adcampaignid and addestination need to be set for Ad Click Attribution to work.
CONSOLE MESSAGE: line 108: Both adcampaignid and addestination need to be set for Ad Click Attribution to work.
CONSOLE MESSAGE: line 108: addestination can not be the same site as the current website.
Test for validity of ad click attribution attributes on anchor tags.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
